+++ app/Http/Controllers/BorrowsController.php
@@ -1,12 +1,18 @@
 <?php
 
 namespace App\Http\Controllers;
-
+use Illuminate\Support\Facades\Validator;
 use Illuminate\Http\Request;
 use App\Borrow;
 
 class BorrowsController extends Controller
 {
+    public function __construct()
+    {
+        $this->middleware('jwt.verify');
+
+    }
+
     /**
      * Display a listing of the resource.
      *
@@ -15,6 +21,10 @@ class BorrowsController extends Controller
     public function index()
     {
         $borrows = Borrow::all();
+        if(!$borrows)
+        {
+            return ["message" => "no borrow"];
+        }
         return $borrows;
     }
 
@@ -36,6 +46,42 @@ class BorrowsController extends Controller
      */
     public function store(Request $request)
     {
+        $validator = Validator::make(
+            [
+                'borrower_id' => $request->borrower_id,
+                'book_id' =>  $request->book_id,
+                'borrowed_at' => $request->borrowed_at,
+                'take_back_at' => $request->take_back_at,
+                'taken_back' => $request->taken_back;
+            ],
+            [
+                'borrower_id' => 'required|integer',
+                'book_id' => 'required|integer',
+                'borrowed_at' => 'required|date_format:Y-m-d:h-m-s',
+                'take_back_at' => 'required|date_format:Y-m-d:h-m-s',
+                'taken_back' => 'boolean'
+            ]
+            );
+        if ($validator->fails())
+        {
+            $error = $validator->errors()->first();
+            return json_encode($error);
+        }
+
+        $user = DB::table('users')->where('id', $request->borrower_id)->first();
+
+        if (!$user)
+        {
+            return ["message" => "borrower dosent exist!"];
+        }
+
+        $book = DB::table('books')->where('id', $request->book_id)->first();
+
+        if (!$book)
+        {
+            return ["message" => "book dosent exist!"];
+        }
+
         $borrow = new Borrow;
         $borrow->borrower_id = $request->borrower_id;
         $borrow->borrowed_at = $request->borrowed_at;
@@ -55,6 +101,12 @@ class BorrowsController extends Controller
     public function show($id)
     {
         $borrow = Borrow::find($id);
+        if (!$borrow)
+        {
+            return ["message" => "no borrow"];
+
+        }
+        retur
         return json_encode($borrow);
     }
 
@@ -79,6 +131,42 @@ class BorrowsController extends Controller
     public function update(Request $request, $id)
     {
         $borrow = Borrow::find($id);
+        $validator = Validator::make(
+            [
+                'borrower_id' => $request->borrower_id,
+                'book_id' =>  $request->book_id,
+                'borrowed_at' => $request->borrowed_at,
+                'take_back_at' => $request->take_back_at,
+                'taken_back' => $request->taken_back;
+            ],
+            [
+                'borrower_id' => 'required|integer',
+                'book_id' => 'required|integer',
+                'borrowed_at' => 'required|date_format:Y-m-d:h-m-s',
+                'take_back_at' => 'required|date_format:Y-m-d:h-m-s',
+                'taken_back' => 'boolean'
+            ]
+            );
+        if ($validator->fails())
+        {
+            $error = $validator->errors()->first();
+            return json_encode($error);
+        }
+
+        $user = DB::table('users')->where('id', $request->borrower_id)->first();
+
+        if (!$user)
+        {
+            return ["message" => "borrower dosent exist!"];
+        }
+
+        $book = DB::table('books')->where('id', $request->book_id)->first();
+
+        if (!$book)
+        {
+            return ["message" => "book dosent exist!"];
+        }
+
         if($borrow)
         {
             $borrow->borrower_id = $request->borrower_id;
@@ -105,7 +193,13 @@ class BorrowsController extends Controller
     public function destroy($id)
     {
         $borrow = Borrow::find($id);
+        if (!$borrow)
+        {
+            return ["message" => "no borrow"];
+
+        }
         $borrow->delete();
         return ["message" => "dropped successfully"];
     }
+
 }
